requirements.txt
request
BeautifulSoup
smtplib
time
MIMEText